+++ rx.h
@@ -93,7 +93,7 @@ class rx
   rx(rx_settings & settings_to_apply, rx_status & status);
   void apply_settings();
   void run();
-  void get_spectrum(float spectrum[], s_filter_control &fc);
+  void get_spectrum(uint8_t spectrum[], s_filter_control &fc);
   rx_settings &settings_to_apply;
   rx_status &status;
   rx_dsp rx_dsp_inst;
